refactor: response 수정 및 guestbook api 수정
