+++ src/main/java/com/kangwon/festival/domain/admin/controller/AdminNoticeController.java
@@ -1,0 +1,4 @@
+package com.kangwon.festival.domain.admin.controller;
+
+public class AdminNoticeController {
+}
