+++ src/main/java/com/kangwon/festival/domain/admin/controller/AdminPostController.java
@@ -1,0 +1,4 @@
+package com.kangwon.festival.domain.admin.controller;
+
+public class AdminPostController {
+}
