+++ src/main/java/com/kangwon/festival/domain/admin/dto/ReportResponse.java
@@ -1,0 +1,16 @@
+package com.kangwon.festival.domain.admin.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+import java.time.LocalDateTime;
+
+@Getter
+@AllArgsConstructor
+public class ReportResponse {
+    private Long reportId;
+    private Long reportedUserId;
+    private Long reporterUserId;
+    private String reason;
+    private LocalDateTime createdDateTime;
+}
