+++ src/main/java/com/kangwon/festival/domain/admin/repository/AdminGuestbookInfoRepository.java
@@ -1,0 +1,8 @@
+package com.kangwon.festival.domain.admin.repository;
+
+import com.kangwon.festival.global.entity.GuestbookInfo;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface AdminGuestbookInfoRepository extends JpaRepository<GuestbookInfo, Integer> {
+
+}
