+++ src/main/java/com/kangwon/festival/domain/admin/repository/AdminJoinboardInfoRepository.java
@@ -1,0 +1,8 @@
+package com.kangwon.festival.domain.admin.repository;
+
+import com.kangwon.festival.global.entity.JoinboardInfo;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface AdminJoinboardInfoRepository extends JpaRepository<JoinboardInfo, Integer> {
+
+}
