+++ src/main/java/com/kangwon/festival/domain/admin/repository/AdminNoticeInfoRepository.java
@@ -1,0 +1,8 @@
+package com.kangwon.festival.domain.admin.repository;
+
+import com.kangwon.festival.global.entity.NoticeInfo;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface AdminNoticeInfoRepository extends JpaRepository<NoticeInfo, Integer> {
+
+}
