+++ src/main/java/com/kangwon/festival/domain/admin/repository/AdminUserInfoRepository.java
@@ -1,0 +1,8 @@
+package com.kangwon.festival.domain.admin.repository;
+
+import com.kangwon.festival.global.entity.UserInfo;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface AdminUserInfoRepository extends JpaRepository<UserInfo, Long> {
+
+}
