+++ src/main/java/com/kangwon/festival/domain/admin/repository/AdminUserReportRepository.java
@@ -1,0 +1,11 @@
+package com.kangwon.festival.domain.admin.repository;
+
+import com.kangwon.festival.global.entity.UserReport;
+import com.kangwon.festival.global.entity.UserInfo;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.List;
+
+public interface AdminUserReportRepository extends JpaRepository<UserReport, Long> {
+    List<UserReport> findByReportedUser(UserInfo reportedUser);
+}
