+++ src/main/java/com/kangwon/festival/domain/admin/service/AdminNoticeService.java
@@ -1,0 +1,4 @@
+package com.kangwon.festival.domain.admin.service;
+
+public class AdminNoticeService {
+}
