+++ src/main/java/com/kangwon/festival/domain/admin/service/AdminPostService.java
@@ -1,0 +1,4 @@
+package com.kangwon.festival.domain.admin.service;
+
+public class AdminPostService {
+}
