+++ src/main/java/com/kangwon/festival/domain/guestbook/dto/GuestbookRequest.java
@@ -1,0 +1,13 @@
+package com.kangwon.festival.domain.guestbook.dto;
+
+import lombok.*;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+public class GuestbookRequest {
+    private Long userId;
+    private String Nickname;
+    private String content;
+}
