+++ src/main/java/com/kangwon/festival/domain/guestbook/repository/GuestbookHistoryRepository.java
@@ -1,0 +1,10 @@
+package com.kangwon.festival.domain.guestbook.repository;
+
+import com.kangwon.festival.global.entity.GuestbookHistory;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface GuestbookHistoryRepository extends JpaRepository<GuestbookHistory, Long> {
+
+}
