+++ src/main/java/com/kangwon/festival/domain/mypage/repository/MypageUserInfoRepository.java
@@ -1,0 +1,11 @@
+package com.kangwon.festival.domain.mypage.repository;
+
+
+import com.kangwon.festival.global.entity.UserInfo;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface MypageUserInfoRepository extends JpaRepository<UserInfo, Long> {
+    Optional<UserInfo> findByUserNickname(String nickname);
+}
