--- src/main/java/com/kangwon/festival/main/admin/controller/AdminPostController.java
@@ -1,4 +1,0 @@
-package com.kangwon.festival.main.admin.controller;
-
-public class AdminPostController {
-}
