--- src/main/java/com/kangwon/festival/main/admin/dto/ReportResponse.java
@@ -1,16 +1,0 @@
-package com.kangwon.festival.main.admin.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-import java.time.LocalDateTime;
-
-@Getter
-@AllArgsConstructor
-public class ReportResponse {
-    private Long reportId;
-    private Long reportedUserId;
-    private Long reporterUserId;
-    private String reason;
-    private LocalDateTime createdDateTime;
-}
