--- src/main/java/com/kangwon/festival/main/admin/repository/AdminGuestbookInfoRepository.java
@@ -1,8 +1,0 @@
-package com.kangwon.festival.main.admin.repository;
-
-import com.kangwon.festival.global.entity.GuestbookInfo;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface AdminGuestbookInfoRepository extends JpaRepository<GuestbookInfo, Integer> {
-
-}
