--- src/main/java/com/kangwon/festival/main/admin/repository/AdminJoinboardInfoRepository.java
@@ -1,8 +1,0 @@
-package com.kangwon.festival.main.admin.repository;
-
-import com.kangwon.festival.global.entity.JoinboardInfo;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface AdminJoinboardInfoRepository extends JpaRepository<JoinboardInfo, Integer> {
-
-}
