--- src/main/java/com/kangwon/festival/main/admin/repository/AdminNoticeInfoRepository.java
@@ -1,8 +1,0 @@
-package com.kangwon.festival.main.admin.repository;
-
-import com.kangwon.festival.global.entity.NoticeInfo;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface AdminNoticeInfoRepository extends JpaRepository<NoticeInfo, Integer> {
-
-}
