--- src/main/java/com/kangwon/festival/main/admin/repository/AdminUserBlockRepository.java
@@ -1,12 +1,0 @@
-package com.kangwon.festival.main.admin.repository;
-
-import com.kangwon.festival.global.entity.UserBlock;
-import com.kangwon.festival.global.entity.UserInfo;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.List;
-
-public interface AdminUserBlockRepository extends JpaRepository<UserBlock, Long> {
-    List<UserBlock> findByBlocker(UserInfo blocker);
-    List<UserBlock> findByBlocked(UserInfo blocked);
-}
