--- src/main/java/com/kangwon/festival/main/admin/service/AdminNoticeService.java
@@ -1,4 +1,0 @@
-package com.kangwon.festival.main.admin.service;
-
-public class AdminNoticeService {
-}
