--- src/main/java/com/kangwon/festival/main/admin/service/AdminPostService.java
@@ -1,4 +1,0 @@
-package com.kangwon.festival.main.admin.service;
-
-public class AdminPostService {
-}
