--- src/main/java/com/kangwon/festival/main/guestbook/dto/GuestbookRequest.java
@@ -1,14 +1,0 @@
-package com.kangwon.festival.main.guestbook.dto;
-
-import lombok.*;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-public class GuestbookRequest {
-    private Long userId;
-    private String title;
-    private String content;
-    private boolean anonymous;
-}
