--- src/main/java/com/kangwon/festival/main/guestbook/service/GuestbookServiceImpl.java
@@ -1,16 +1,0 @@
-package com.kangwon.festival.main.guestbook.service;
-
-public interface GuestbookServiceImpl {
-
-    // @MethodDescription(description = "방명록 등록")
-
-    // @MethodDescription(description = "방명록 조회")
-
-    // @MethodDescription(description = "내 방명록 조회")
-
-    // @MethodDescription(description = "내가 쓴 방명록 수정")
-
-    // @MethodDescription(description = "방명록 삭제")
-
-
-}
